src/requirements.txt
streamlit
snowflake-connector-python
snowflake-snowpark-python
flask
pandas